server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
googlemaps==4.10.0
requests==2.32.3
langchain==0.3.0
langchain-google-genai==2.0.0
pydantic==2.9.2
